Tests/LibWeb/Text/expected/abspos-relayout-move-multiple-boxes.txt

a=0,100,40,30
b=50,150,40,30
c=200,0,40,30
partial-layouts=0 full-layouts=1
FAIL
